Refactor ir package a bit
